+++ cmd/upgrade.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/Azure/acs-engine/pkg/acsengine"
 	"github.com/Azure/acs-engine/pkg/api"
-	"github.com/Azure/acs-engine/pkg/operations/armhelpers"
+	"github.com/Azure/acs-engine/pkg/armhelpers"
 
 	"github.com/Azure/acs-engine/pkg/operations"
 	"github.com/Azure/go-autorest/autorest/adal"
